% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icd_desc.R
\name{add_icd_desc}
\alias{add_icd_desc}
\title{Add ICD-10 descriptions to a data frame}
\usage{
add_icd_desc(df, code_col = "code")
}
\arguments{
\item{df}{A data frame containing ICD codes.}

\item{code_col}{Column name with ICD codes.}
}
\value{
The input data frame with a new ICD_desc column.
}
\description{
Add ICD-10 descriptions to a data frame
}
