% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icd_desc_any.R
\name{add_icd_desc_any}
\alias{add_icd_desc_any}
\title{Add ICD-10 descriptions to a vector of codes}
\usage{
add_icd_desc_any(x)
}
\arguments{
\item{x}{A character vector of ICD codes.}
}
\value{
A character vector of descriptions.
}
\description{
Add ICD-10 descriptions to a vector of codes
}
