% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icd_desc_df.R
\name{add_icd_desc_df}
\alias{add_icd_desc_df}
\title{Add ICD descriptions next to code column}
\usage{
add_icd_desc_df(df, code_col = "code", desc_col = "desc")
}
\arguments{
\item{df}{A data frame.}

\item{code_col}{Column name containing ICD codes, given as a string.}

\item{desc_col}{Name of the new description column.}
}
\value{
The input data frame with description column inserted after code_col.
}
\description{
Takes a data frame with a code column and inserts a description column
immediately after it.
}
