% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proc_desc_any.R
\name{add_proc_desc_any}
\alias{add_proc_desc_any}
\title{Add procedure descriptions to a vector of codes}
\usage{
add_proc_desc_any(x)
}
\arguments{
\item{x}{A character vector of procedure codes.}
}
\value{
A character vector of descriptions.
}
\description{
Add procedure descriptions to a vector of codes
}
