% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_proc_desc_df.R
\name{add_proc_desc_df}
\alias{add_proc_desc_df}
\title{Add ICD-10-PCS descriptions to a data frame}
\usage{
add_proc_desc_df(df, proc_col = "proc", desc_col = "desc")
}
\arguments{
\item{df}{A data frame.}

\item{proc_col}{Column name containing procedure codes, given as a string.}

\item{desc_col}{Name of the new description column.}
}
\value{
The input data frame with description column inserted after proc_col.
}
\description{
Takes a data frame with a procedure code column and inserts a description
column immediately after it.
}
