% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rule_desc.R
\name{add_rule_desc}
\alias{add_rule_desc}
\title{Add descriptions to rule table codes}
\usage{
add_rule_desc(rules_df)
}
\arguments{
\item{rules_df}{A data frame with LHS and RHS columns.}
}
\value{
The same data frame with LHS_desc and RHS_desc added.
}
\description{
Add descriptions to rule table codes
}
