% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_icd_dictionary.R
\name{load_icd_dictionary}
\alias{load_icd_dictionary}
\title{Load ICD-10 dictionary}
\usage{
load_icd_dictionary()
}
\value{
A data frame with ICD-10 codes and descriptions.
}
\description{
Reads the ICD-10 code dictionary bundled inside the package.
}
