% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_desc.R
\name{proc_desc}
\alias{proc_desc}
\title{Get ICD-10-PCS description for a procedure code}
\usage{
proc_desc(code)
}
\arguments{
\item{code}{Procedure code(s), quoted or unquoted.}
}
\value{
A character vector of descriptions.
}
\description{
Looks up the description for one or more ICD-10-PCS procedure codes using
the package procedure dictionary.
}
