% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_rules_with_desc.R
\name{proc_rules_with_desc}
\alias{proc_rules_with_desc}
\title{Add procedure descriptions to association rules}
\usage{
proc_rules_with_desc(rules)
}
\arguments{
\item{rules}{An object of class rules from arules.}
}
\value{
A data.frame with descriptions placed next to lhs and rhs.
}
\description{
Add procedure descriptions to association rules
}
