% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_with_desc.R
\name{rules_with_desc}
\alias{rules_with_desc}
\title{Add ICD descriptions to association rules}
\usage{
rules_with_desc(rules)
}
\arguments{
\item{rules}{An object of class rules from arules.}
}
\value{
A data.frame with descriptions placed next to lhs and rhs.
}
\description{
Add ICD descriptions to association rules
}
